chore: Rename module path.
